Implemented PUT method. Added tests.

+++ src/API/Write/Controllers/v1/UsersControllers.cs
@@ -4,7 +4,7 @@
 [ApiVersion("1.0")]
 [Route("api/v{version:apiVersion}/users")]
 [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-public class UsersController : API.Shared.Controllers.v1.BaseController
+public class UsersController : Users.API.Shared.Controllers.v1.BaseController
 {
     private readonly ILogger<UsersController> _logger;
     private readonly IMediator _mediator;
@@ -19,10 +19,6 @@ public class UsersController : API.Shared.Controllers.v1.BaseController
         _mediator = mediator;
         _mapper = mapper;
     }
-
-    /***********************************************************************************************************************
-     * POST --> ADD                                                                                                     *
-     ***********************************************************************************************************************/                                                                                                                     
 
     /// <summary>
     /// Add a user
@@ -41,20 +37,14 @@ public class UsersController : API.Shared.Controllers.v1.BaseController
         if (addUserRequest == null) return BadRequest();
 
         var addUserCommand = _mapper.Map<Users.Application.Commands.AddUserCommand>(addUserRequest);
-
         var user = await _mediator.Send(addUserCommand);
-
         var userResponse = _mapper.Map<Users.API.Models.Response.v1.UserResponse>(user);
 
         return Created($"http://localhost/api/v1/users/{userResponse.Id}", userResponse);
     }
 
-    /***********************************************************************************************************************
-     * DELETE                                                                                                              *
-     ***********************************************************************************************************************/
-
     /// <summary>
-    /// Delete a user
+    /// Delete an existing user
     /// </summary>
     /// <param name="userId">The unique identifier of the user</param>
     /// <returns>Nothing</returns>
@@ -69,6 +59,8 @@ public class UsersController : API.Shared.Controllers.v1.BaseController
     {
         if (userId == default) return BadRequest();
 
+        if (!await DoesUserExist(userId)) return NotFound();
+
         var deleteUserRequest = new Users.API.Models.Request.v1.DeleteUserRequest(userId);
 
         var deleteUserCommand = _mapper.Map<Users.Application.Commands.DeleteUserCommand>(deleteUserRequest);
@@ -78,64 +70,53 @@ public class UsersController : API.Shared.Controllers.v1.BaseController
         return NoContent();
     }
 
-    /***********************************************************************************************************************
-     * PUT -- UPSERT                                                                                                       *
-     ***********************************************************************************************************************/
-
     /// <summary>
-    /// Upsert a user.
+    /// Update an existing user.
     /// </summary>
     /// <param name="userId">The unique identifier of the user</param>
-    /// <param name="upsertUserRequest">
-    /// An UpsertUserRequest object which contains all the
-    /// data required to either update or create a user.
-    /// </param>
-    /// <returns>The upserted user</returns>
+    /// <param name="updateUserRequest">An UpdateUserRequest object which contains the data required to update a user.</param>
+    /// <returns>The updated user</returns>
     /// <response code="200">Success - OK - The user was successfully updated</response>
-    /// <response code="201">Success - Created - The user was successfully created</response>
-    /// <response code="400">Error - Bad Request - It was not possible to bind the request JSON</response>
+    /// <response code="404">Error - Not Found - Could not find the corresponding user for the given identifier</response>
+    /// <response code="422">Error - Unprocessable Entity - It was not possible to bind the request JSON</response>
     [HttpPut("{userId:guid}")]
     [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status201Created)]
-    [ProducesResponseType(StatusCodes.Status400BadRequest)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
 
     public async Task<IActionResult> UpsertUser(
         [FromRoute] Guid userId,
-        [FromBody] Users.API.Models.Request.v1.UpsertUserRequest upsertUserRequest)
+        [FromBody] Users.API.Models.Request.v1.UpdateUserRequest updateUserRequest)
     {
-        if (upsertUserRequest == null) return BadRequest();
+        if (updateUserRequest == null) return BadRequest();
 
         if (await DoesUserExist(userId))
         {
-            var updateUserCommand = _mapper.Map<Application.Commands.UpdateUserCommand>(upsertUserRequest);
+            var updateUserCommand = _mapper.Map<Application.Commands.UpdateUserCommand>(updateUserRequest);
             updateUserCommand.Id = userId;
-
             var updatedUser = await _mediator.Send(updateUserCommand);
 
-            return Ok();
+            return Ok(updatedUser);
         }
 
-        var createUserRequest = new Users.API.Models.Request.v1.AddUserRequest();
-        return RedirectToAction(nameof(AddUser), createUserRequest);
+        var addUserCommand = _mapper.Map<Users.Application.Commands.AddUserCommand>(updateUserRequest);
+        addUserCommand.Id = userId;
+        var user = await _mediator.Send(addUserCommand);
+        var userResponse = _mapper.Map<Users.API.Models.Response.v1.UserResponse>(user);
+
+        return Created($"http://localhost/api/v1/users/{userResponse.Id}", userResponse);
     }
 
-    /***********************************************************************************************************************
-     * PATCH --> UPSERT                                                                                                    *
-     ***********************************************************************************************************************/
-
     /// <summary>
-    /// Fully or partially update a user
+    /// Fully or partially update an existing user
     /// </summary>
     /// <param name="userId">The unique identifier of the user</param>
-    /// <param name="patchDocument">
-    /// A JSON Patch Document detailing the full or partial updates to the user
-    /// </param>
+    /// <param name="patchDocument">A JSON Patch Document detailing the full or partial updates to the user</param>
     /// <returns>The updated user</returns>
     /// <response code="200">Success - OK - The user was successfully updated</response>
     /// <response code="400">Error - Bad Request - It was not possible to bind the request JSON</response>
-    /// <response code="404">Error - Not Found - No user matched the given identifier</response>
-    /// <response code="422">Error - Unprocessable Entity - Unable to process the contained instructions.</response>
+    /// <response code="404">Error - Not Found - Could not find the corresponding user for the given identifier</response>
+    /// <response code="422">Error - Unprocessable Entity - Unable to process the contained instructions</response>
     /// <remarks>
     /// Sample request (this request updates the user's first name): \
     /// PATCH /users/id \
@@ -152,22 +133,22 @@ public class UsersController : API.Shared.Controllers.v1.BaseController
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     [ProducesResponseType(StatusCodes.Status422UnprocessableEntity)]
-    public async Task<IActionResult> UpsertUser(
+    public async Task<IActionResult> PartiallyUser(
         [FromRoute] Guid userId,
-        [FromBody] JsonPatchDocument<Users.API.Models.Request.v1.PartiallyUpsertUserRequest> patchDocument)
+        [FromBody] JsonPatchDocument<Users.API.Models.Request.v1.PartiallyUpdateUserRequest> patchDocument)
     {
         if (patchDocument == null) return BadRequest();
 
-        var partiallyUpsertUserRequest = new Users.API.Models.Request.v1.PartiallyUpsertUserRequest(userId);
-        patchDocument.ApplyTo(partiallyUpsertUserRequest, ModelState);
-        if (!TryValidateModel(partiallyUpsertUserRequest))
+        var partiallyUpdateUserRequest = new Users.API.Models.Request.v1.PartiallyUpdateUserRequest(userId);
+        patchDocument.ApplyTo(partiallyUpdateUserRequest, ModelState);
+        if (!TryValidateModel(partiallyUpdateUserRequest))
         {
             return ValidationProblem(ModelState);
         }
             
         if (await DoesUserExist(userId))
         {
-            var updatedUser = await PartiallyUpdateUser(userId, partiallyUpsertUserRequest);
+            var updatedUser = await PartiallyUpdateUser(userId, partiallyUpdateUserRequest);
             return Ok(updatedUser);
         }
             
@@ -197,23 +178,11 @@ public class UsersController : API.Shared.Controllers.v1.BaseController
         return doesUserExist;
     }
         
-    //private async Task<Users.API.Models.Response.v1.UserResponse> UpdateUser(
-    //    Guid userId,
-    //    Users.API.Models.Request.v1.UpsertUserRequest request)
-    //{
-    //    if (request == null) throw new ArgumentNullException(nameof(request));
-
-    //    // Update User...
-    //    await Task.CompletedTask;
-            
-    //    throw new NotImplementedException();
-    //}
-
     private async Task<Users.API.Models.Response.v1.UserResponse> PartiallyUpdateUser(
         Guid userId,
-        Users.API.Models.Request.v1.PartiallyUpsertUserRequest partiallyUpsertUserRequest)
+        Users.API.Models.Request.v1.PartiallyUpdateUserRequest partiallyUpdateUserRequest)
     {
-        if (partiallyUpsertUserRequest == null) throw new ArgumentNullException(nameof(partiallyUpsertUserRequest));
+        if (partiallyUpdateUserRequest == null) throw new ArgumentNullException(nameof(partiallyUpdateUserRequest));
             
         // var partiallyUpdateUserCommand = _mapper.Map<PartiallyUpdateUserCommand>(partiallyUpsertUserRequest);
         //
